# libroach: optimize MVCC reverse scans