@petermattis 2018-01-13
I think making iters-per-seek adaptive might address the performance hit for the `versions=100` benchmarks. I'm going to experiment with that soon.