@petermattis 2018-01-13
I added a second commit which makes iters-before-seek adaptive. @arjunravinarayan please take a look.