@petermattis 2018-01-13
Keeping track of all of the benchmark numbers and what versions they apply to is a bit daunting. The following numbers are using `old` from before the reimplementation work began and `new` using this PR:

```
name                                                          old time/op    new time/op     delta
MVCCScan_RocksDB/rows=1/versions=1/valueSize=8-8                6.02µs ± 1%     5.30µs ± 2%   -12.07%  (p=0.000 n=8+10)
MVCCScan_RocksDB/rows=1/versions=2/valueSize=8-8                6.93µs ± 2%     5.43µs ± 5%   -21.64%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1/versions=10/valueSize=8-8               8.41µs ± 3%     5.85µs ± 2%   -30.37%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1/versions=100/valueSize=8-8              9.48µs ± 2%     7.04µs ± 3%   -25.74%  (p=0.000 n=9+10)
MVCCScan_RocksDB/rows=10/versions=1/valueSize=8-8               12.7µs ± 2%      8.3µs ± 1%   -34.75%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=10/versions=2/valueSize=8-8               20.0µs ± 1%      9.0µs ± 1%   -55.25%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=10/versions=10/valueSize=8-8              31.3µs ± 2%     14.3µs ± 1%   -54.43%  (p=0.000 n=10+9)
MVCCScan_RocksDB/rows=10/versions=100/valueSize=8-8             38.2µs ± 2%     30.9µs ± 4%   -19.19%  (p=0.000 n=9+10)
MVCCScan_RocksDB/rows=100/versions=1/valueSize=8-8              72.6µs ± 1%     31.0µs ± 2%   -57.38%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=2/valueSize=8-8               142µs ± 2%       37µs ± 1%   -73.73%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=10/valueSize=8-8              246µs ± 2%       81µs ± 2%   -67.20%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=100/valueSize=8-8             304µs ± 1%      241µs ± 2%   -20.92%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=1/valueSize=8-8              635µs ± 2%      246µs ± 1%   -61.20%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=2/valueSize=8-8             1.33ms ± 2%     0.30ms ± 1%   -77.30%  (p=0.000 n=10+9)
MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-8            2.35ms ± 2%     0.71ms ± 1%   -69.62%  (p=0.000 n=9+10)
MVCCScan_RocksDB/rows=1000/versions=100/valueSize=8-8           2.94ms ± 2%     2.29ms ± 1%   -22.08%  (p=0.000 n=10+9)
MVCCReverseScan_RocksDB/rows=1/versions=1/valueSize=8-8         10.9µs ± 1%      7.2µs ± 2%   -34.23%  (p=0.000 n=9+9)
MVCCReverseScan_RocksDB/rows=1/versions=2/valueSize=8-8         12.0µs ± 2%      8.1µs ± 3%   -32.34%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=1/versions=10/valueSize=8-8        12.9µs ± 1%     11.3µs ± 2%   -12.57%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=1/versions=100/valueSize=8-8       14.1µs ± 1%     15.0µs ± 2%    +6.22%  (p=0.000 n=10+8)
MVCCReverseScan_RocksDB/rows=10/versions=1/valueSize=8-8        34.9µs ± 1%     11.2µs ± 2%   -67.86%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=10/versions=2/valueSize=8-8        43.3µs ± 1%     15.7µs ± 1%   -63.82%  (p=0.000 n=9+8)
MVCCReverseScan_RocksDB/rows=10/versions=10/valueSize=8-8       51.5µs ± 1%     35.6µs ± 5%   -30.90%  (p=0.000 n=9+10)
MVCCReverseScan_RocksDB/rows=10/versions=100/valueSize=8-8      57.8µs ± 1%     51.3µs ± 2%   -11.22%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=1/valueSize=8-8        264µs ± 3%       44µs ± 2%   -83.35%  (p=0.000 n=10+9)
MVCCReverseScan_RocksDB/rows=100/versions=2/valueSize=8-8        342µs ± 1%       82µs ± 1%   -75.99%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=10/valueSize=8-8       417µs ± 2%      237µs ± 2%   -43.13%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=100/valueSize=8-8      482µs ± 2%      391µs ± 2%   -18.84%  (p=0.000 n=10+8)
MVCCReverseScan_RocksDB/rows=1000/versions=1/valueSize=8-8      2.52ms ± 2%     0.36ms ± 2%   -85.74%  (p=0.000 n=9+10)
MVCCReverseScan_RocksDB/rows=1000/versions=2/valueSize=8-8      3.28ms ± 2%     0.73ms ± 1%   -77.64%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=1000/versions=10/valueSize=8-8     4.08ms ± 3%     2.22ms ± 2%   -45.67%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=1000/versions=100/valueSize=8-8    4.66ms ± 1%     3.76ms ± 3%   -19.14%  (p=0.000 n=7+10)

name                                                          old speed      new speed       delta
MVCCScan_RocksDB/rows=1/versions=1/valueSize=8-8              1.33MB/s ± 1%   1.51MB/s ± 2%   +13.85%  (p=0.000 n=8+10)
MVCCScan_RocksDB/rows=1/versions=2/valueSize=8-8              1.16MB/s ± 2%   1.48MB/s ± 5%   +27.71%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1/versions=10/valueSize=8-8              956kB/s ± 2%   1366kB/s ± 2%   +42.95%  (p=0.000 n=9+10)
MVCCScan_RocksDB/rows=1/versions=100/valueSize=8-8             843kB/s ± 2%   1137kB/s ± 2%   +34.82%  (p=0.000 n=9+10)
MVCCScan_RocksDB/rows=10/versions=1/valueSize=8-8             6.31MB/s ± 2%   9.67MB/s ± 1%   +53.26%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=10/versions=2/valueSize=8-8             4.00MB/s ± 1%   8.93MB/s ± 1%  +123.09%  (p=0.000 n=9+10)
MVCCScan_RocksDB/rows=10/versions=10/valueSize=8-8            2.56MB/s ± 2%   5.61MB/s ± 1%  +119.40%  (p=0.000 n=10+9)
MVCCScan_RocksDB/rows=10/versions=100/valueSize=8-8           2.08MB/s ± 5%   2.59MB/s ± 4%   +24.46%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=1/valueSize=8-8            11.0MB/s ± 1%   25.8MB/s ± 2%  +134.63%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=2/valueSize=8-8            5.65MB/s ± 2%  21.52MB/s ± 1%  +280.65%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=10/valueSize=8-8           3.25MB/s ± 2%   9.92MB/s ± 2%  +204.92%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=100/valueSize=8-8          2.63MB/s ± 1%   3.33MB/s ± 1%   +26.46%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=1/valueSize=8-8           12.6MB/s ± 2%   32.5MB/s ± 1%  +157.74%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=2/valueSize=8-8           6.03MB/s ± 2%  26.56MB/s ± 1%  +340.39%  (p=0.000 n=10+9)
MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-8          3.41MB/s ± 1%  11.22MB/s ± 1%  +229.06%  (p=0.000 n=9+10)
MVCCScan_RocksDB/rows=1000/versions=100/valueSize=8-8         2.72MB/s ± 2%   3.48MB/s ± 1%   +28.13%  (p=0.000 n=10+8)
MVCCReverseScan_RocksDB/rows=1/versions=1/valueSize=8-8        730kB/s ± 0%   1113kB/s ± 2%   +52.51%  (p=0.000 n=6+9)
MVCCReverseScan_RocksDB/rows=1/versions=2/valueSize=8-8        667kB/s ± 2%    985kB/s ± 3%   +47.68%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=1/versions=10/valueSize=8-8       618kB/s ± 2%    706kB/s ± 2%   +14.24%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=1/versions=100/valueSize=8-8      565kB/s ± 1%    533kB/s ± 3%    -5.60%  (p=0.000 n=10+9)
MVCCReverseScan_RocksDB/rows=10/versions=1/valueSize=8-8      2.29MB/s ± 1%   7.13MB/s ± 2%  +210.90%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=10/versions=2/valueSize=8-8      1.84MB/s ± 1%   5.11MB/s ± 1%  +176.95%  (p=0.000 n=8+8)
MVCCReverseScan_RocksDB/rows=10/versions=10/valueSize=8-8     1.55MB/s ± 1%   2.25MB/s ± 5%   +44.63%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=10/versions=100/valueSize=8-8    1.38MB/s ± 1%   1.56MB/s ± 2%   +12.64%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=1/valueSize=8-8     3.03MB/s ± 3%  18.13MB/s ± 4%  +498.05%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=2/valueSize=8-8     2.34MB/s ± 1%   9.74MB/s ± 1%  +316.55%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=10/valueSize=8-8    1.92MB/s ± 2%   3.37MB/s ± 2%   +75.81%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=100/valueSize=8-8   1.66MB/s ± 2%   2.04MB/s ± 2%   +23.19%  (p=0.000 n=10+8)
MVCCReverseScan_RocksDB/rows=1000/versions=1/valueSize=8-8    3.17MB/s ± 2%  22.26MB/s ± 2%  +601.23%  (p=0.000 n=9+10)
MVCCReverseScan_RocksDB/rows=1000/versions=2/valueSize=8-8    2.44MB/s ± 2%  10.90MB/s ± 1%  +347.07%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=1000/versions=10/valueSize=8-8   1.96MB/s ± 3%   3.61MB/s ± 2%   +83.91%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=1000/versions=100/valueSize=8-8  1.72MB/s ± 1%   2.12MB/s ± 3%   +23.65%  (p=0.000 n=7+10)
```

There is only one benchmark that shows a performance hit: `MVCCReverseScan_RocksDB/rows=1/versions=100/valueSize=8`. The reason for the hit is that previously we were doing 2 `iter->Seek` calls and now we're doing 2 `iter->Seek` + 5 `iter->Prev` (this is the `kMaxItersBeforeSeek` optimization). If we tweak that optimization we can get rid of the performance hit on this benchmark, but it would bring down the numbers for other benchmarks.