@petermattis 2018-01-13
@arjunravinarayan Your suggestions wins in some benchmarks and loses in others. And some of the losses are quite dramatic. (`old` is this PR, `new` is this PR plus your heuristic).

```
name                                                          old time/op    new time/op    delta
MVCCScan_RocksDB/rows=1/versions=1/valueSize=8-8                5.50µs ± 2%    5.46µs ± 1%     ~     (p=0.211 n=9+10)
MVCCScan_RocksDB/rows=1/versions=2/valueSize=8-8                5.54µs ± 2%    5.76µs ± 4%   +4.08%  (p=0.005 n=9+10)
MVCCScan_RocksDB/rows=1/versions=10/valueSize=8-8               6.06µs ± 2%    6.15µs ± 2%   +1.48%  (p=0.001 n=9+10)
MVCCScan_RocksDB/rows=1/versions=100/valueSize=8-8              7.31µs ± 3%    7.47µs ± 3%   +2.20%  (p=0.003 n=10+10)
MVCCScan_RocksDB/rows=10/versions=1/valueSize=8-8               8.48µs ± 1%    8.50µs ± 1%     ~     (p=0.516 n=10+10)
MVCCScan_RocksDB/rows=10/versions=2/valueSize=8-8               9.30µs ± 2%    9.31µs ± 2%     ~     (p=0.670 n=10+10)
MVCCScan_RocksDB/rows=10/versions=10/valueSize=8-8              14.8µs ± 2%    15.4µs ± 2%   +3.94%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=10/versions=100/valueSize=8-8             32.3µs ± 2%    32.0µs ± 3%     ~     (p=0.133 n=9+10)
MVCCScan_RocksDB/rows=100/versions=1/valueSize=8-8              31.7µs ± 1%    32.5µs ± 6%   +2.72%  (p=0.000 n=10+10)
MVCCScan_RocksDB/rows=100/versions=2/valueSize=8-8              38.0µs ± 1%    38.6µs ± 1%   +1.66%  (p=0.000 n=10+9)
MVCCScan_RocksDB/rows=100/versions=10/valueSize=8-8             83.4µs ± 1%    83.2µs ± 2%     ~     (p=0.842 n=9+10)
MVCCScan_RocksDB/rows=100/versions=100/valueSize=8-8             254µs ± 3%     249µs ± 2%   -2.08%  (p=0.005 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=1/valueSize=8-8              248µs ± 1%     250µs ± 2%     ~     (p=0.165 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=2/valueSize=8-8              308µs ± 2%     310µs ± 1%     ~     (p=0.079 n=10+9)
MVCCScan_RocksDB/rows=1000/versions=10/valueSize=8-8             734µs ± 2%     728µs ± 1%     ~     (p=0.247 n=10+10)
MVCCScan_RocksDB/rows=1000/versions=100/valueSize=8-8           2.44ms ± 2%    2.40ms ± 2%   -1.49%  (p=0.043 n=10+10)
MVCCReverseScan_RocksDB/rows=1/versions=1/valueSize=8-8         7.44µs ± 2%    7.56µs ± 3%     ~     (p=0.055 n=10+10)
MVCCReverseScan_RocksDB/rows=1/versions=2/valueSize=8-8         8.31µs ± 2%    8.34µs ± 2%     ~     (p=0.516 n=10+10)
MVCCReverseScan_RocksDB/rows=1/versions=10/valueSize=8-8        11.8µs ± 2%    11.8µs ± 2%     ~     (p=0.858 n=9+10)
MVCCReverseScan_RocksDB/rows=1/versions=100/valueSize=8-8       15.6µs ± 3%    15.1µs ± 3%   -3.57%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=10/versions=1/valueSize=8-8        11.6µs ± 3%    11.6µs ± 1%     ~     (p=0.353 n=10+10)
MVCCReverseScan_RocksDB/rows=10/versions=2/valueSize=8-8        16.3µs ± 2%    16.3µs ± 3%     ~     (p=0.781 n=10+10)
MVCCReverseScan_RocksDB/rows=10/versions=10/valueSize=8-8       36.1µs ± 2%    39.0µs ± 2%   +7.92%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=10/versions=100/valueSize=8-8      53.7µs ± 2%    51.3µs ± 3%   -4.41%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=1/valueSize=8-8       45.5µs ± 1%    45.1µs ± 2%   -0.98%  (p=0.008 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=2/valueSize=8-8       85.8µs ± 2%    84.8µs ± 1%   -1.26%  (p=0.035 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=10/valueSize=8-8       251µs ± 1%     293µs ± 1%  +16.93%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=100/versions=100/valueSize=8-8      409µs ± 3%     396µs ± 2%   -3.20%  (p=0.000 n=9+10)
MVCCReverseScan_RocksDB/rows=1000/versions=1/valueSize=8-8       370µs ± 1%     364µs ± 2%   -1.50%  (p=0.001 n=9+10)
MVCCReverseScan_RocksDB/rows=1000/versions=2/valueSize=8-8       760µs ± 2%     752µs ± 2%     ~     (p=0.075 n=10+10)
MVCCReverseScan_RocksDB/rows=1000/versions=10/valueSize=8-8     2.32ms ± 2%    2.85ms ± 3%  +22.64%  (p=0.000 n=10+10)
MVCCReverseScan_RocksDB/rows=1000/versions=100/valueSize=8-8    3.93ms ± 3%    3.94ms ± 8%     ~     (p=0.393 n=10+10)
```

Futzing with the adaptation heuristic is likely only going to give small wins. The big win for speeding up scans with lots of versions is to move the historic versions somewhere else.